\name{NTsets-package}
\alias{NTsets-package}
\alias{NTsets}
\docType{package}
\title{
Normality Test Sets Integrated Function and Figure Methods for Continuous Variables
}
\description{
Conduct normality test with nine different function methods and three different figure methods for continuous variables with a friendly GUI. For different data types, provide a more comprehensive means of judgement.
}
\details{
The DESCRIPTION file:
Users can use library(NTsets) to start the GUI and use 'NT()' to restart the programe.
}
\author{
Ren Wen-Long,Xiao Jing,Zhang Ya-Wen,Liang Zhi-Kai
Maintainer:Wenlong Ren<wenlongren@ntu.edu.cn>
}

\keyword{
normality test
 }

\examples{
library(NTsets)
}
